Tweaked the Advanced Settings activity to remove UI elements from the layout rather than just hide them if the element isn't needed. This is primarily used to show/hide the different elements related to selecting file managers. If there are no file managers installed, a TextView tells the user which ones they can install; if file managers are found, a Spinner lets them pick one. Previously, I just hid the unneeded element, which left an ugly space if the Spinner was hidden but the "none found" label was show.
Added support specifically for Android 3.0 (Honeycomb). Cryptnos now uses the new "holographic" theme if it runs on a 3.0 device. Also added dedicated layouts for "extra large" screens (i.e. Honeycomb tablets); the default UI is too small and wastes too much space on these screens, so the special large layouts make it look a lot nicer.

There is a known issue with the Import and Export activities with Honeycomb and when a third-party file manager is selected. After the file manager sends back the selected file/folder, the Cryptnos activity crashes, requiring a Force Close. This does not occur if no file manager is installed or no preference is set; it also does not appear to occur in Android versions older than Honeycomb. So far, I can't trace back the source of the error, but it has something to do with reusing a Cursor that has already been closed. (There should be no cursors involved in these activities.) I'm still investigating, but I'm putting up the new code to the repository anyway.