Minor UI enhancements. One design choice I made early on was that Cryptnos activities should not store state, forcing the user to go back to the main menu if the user left Cryptnos, used another application for a while, then came back. The motivation for this was primarily security; I didn't want someone other than the user to open up Cryptnos and see what the last password generated might be.
The caveat to this, however, is that the user will also be forced back to the main menu if they do something simple like rotate the screen or slide out a physical keyboard. At the least, password and parameter fields get emptied out, which can be a minor inconvenience. This happens because Android destroys and rebuilds activities when minor "configuration changes" like these occur. To preserve the current state, the developer is supposed to store the instance state of the activity and restore it in onCreate(). However, as stated above, we don't do that intentionally.

The answer is to use Activity.onRetainNonConfigurationInstance() to store a temporary object which can store our data, then restore it in onCreate() using Activity.getLastNonConfigurationInstance(). This isn't as persistent as traditionally saving instance state, but it gets us around the short-term issue of small configuration changes.

I also tentatively bumped the version number to 1.2.2. I'll be testing this for a while on my device and in the emulator before pushing it out to the Market. I may sit on this change for a bit until I find something else that needs adding or fixing, as this seems pretty trivial to push out for such a small change.