Fixed a minor bug in the ImportExportHandler. Not sure if this was introduced with the fix for Issue #3 or if it has existed all long. When I tested the updated code for Issue #3, I went back and installed a copy of Cryptnos 1.0 to test the import of the old export format. I exported some sites from v1.0, wiped the device, then installed the new 1.2.3 code. When I tried to import the file, it produced an error stating that the file was invalid. However, when I went back to the main menu, I saw that all the sites in the file were imported successfully.
After finding a copy of the old 1.0 code, I discovered that the original exporter added an extra, empty line at the end of the file before encrypting it. I'm not sure when this problem was introduced, but the 1.2.3 code (well, post-1.2.2) read this extra line as another site parameter line. When it hit this line and attempted to parse it, it threw an error, stating the file was invalid. However, by this point all the valid sites were already imported and existed in the database.

The code now skips the last line of imported data and adjusts all its calculations and messages accordingly. The cross-platform export format (introduced in 1.1) is unaffected by this issue.