Initial import of version 1.3.0 changes:
Added QRCodeHandler class, which does the heavy lifting of importing and exporting parameters via QR code. Like FileManager, this abstracts many possible third-party apps into a single class the rest of the app can call. CryptnosApplication will store the common QRCodeHandler object for the entire app.

Modified CryptnosMainMenu to handle the bulk of the front-end for importing vi QR code, and SiteListActivity to handle exporting to QR code.

Added QR code scanner preference to AdvancedSettingsActivity.

Minor tweaks and improvements to a number of other classes.

This makes the QR import/export code largely complete. ZXing Barcode Scanner is fully supported, but QRDroid support is incomplete and untested. Still need to decide if we'll support any other scanners. There is a known issue with a FC during orientation change on the main menu after a successful scan, but I haven't tracked this down yet because it's hard to debug (I can't scan QR codes in the emulators so I can trace easily) and difficult to replicate.

There's one more major feature that needs to be added for 1.3.0, which is selective imports from a file. That will be the next big step once QR code support is stable.