Fixed bug in preferred QR code drop-down in Advanced Settings which prevented the preference from actually being saved. This was obscured when only one third-party app was installed but became apparent when multiple apps were available.
Fixed orientation change bug after a QR code was scanned, the system determined that the incoming site parameters would overwrite an existing one in the database, and the user said this was OK.

Started implementation of QR Droid and QR Droid Private support. Import works fine. Export is still a bit rough and requires a bit of tweaking in the internal QR code viewer activity. These should be ironed out soon. Had to make several changes to QRCodeHandler to support this, which should improve abstraction when support for other QR code apps is added.