Fixed ImportActivity handing of configuration changes (orientation change, sliding out physical keyboard, etc.)
Swapped default for master password fields back to password mode, then remove the transformation from the field if the Show Master Passwords option is set. In the original (and opposite) implementation, we added the password transformation when it wasn't there, but this caused certain on-screen keyboards to display word completion prompts, resulting in the user's password being displayed on screen whether they wanted it to be or not. Reversing this by making the box default to password mode and then removing it by setting the transformation to null forces the word completion prompts to go away. Word completion won't show up for passwords when the user chooses to show the password, but that's safer than showing the password in word complete prompts when the user doesn't want it displayed.

Moved the About menu option off the main menu and into the Help menu