Tweaked a few more items with respect to the new icon-based menu. I'm having trouble with the "extra high DPI" resources, which one virtual machine the emulator insists on displaying incorrectly. Sadly, I can't find a good way to debug this. It may have to wait until after 1.3.0 is released.
Added "Show Debug Info" checkbox and view to Settings activity. This provides a little bit of debugging info which the user can toggle on and off. This was originally added to help debug the resource issue mentioned above, but I thought it might be handy to leave in so I expanded it. The checkbox and visibility of the text box do not handle screen rotations gracefully, but I don't consider this to be critical. I think it's a bug in Android itself as the checkbox appears to be checked but reports that it isn't.