Added a progress dialog to the database insert portion of the import (from a file) activity. I found that in practice this took way to long to do in the UI thread, especially if there were a lot of sites to import. (I discovered this when I completely uninstalled and reinstalled Cryptnos and had to restore my sites from backup.)
Had to remove and add back in the XHDPI (96 pixel) Settings icon. I tweaked the color at one point to make sure I was using the right icon, but when I tried to change it back Eclipse refused to add the changed file back. I had to actually delete the tweaked file and add the old file back under a new name to get it to take.