Implemented Issue 8 with an enhancement to mimic the behavior of the Windows client by adding an option to clear the master and generated password text boxes whenever Cryptnos goes into the background, such as when the user launches another app. By default this option is turned off to replicate past behavior. This was a bit tricky because it interferes with the code to preserve these boxes when the user changes configurations (rotates the screen, slides out a keyboard, etc.), but it was possible to have our cake and eat it as well in this instance.
Also implemented Issue 11, another enhancement cloning a Windows client feature. Now if the user presses the Enter key (either a physical key or an on-screen key) in the master password box, the events of the Generate button are triggered as if the user tapped the button. This is exactly what the Windows client did, so it makes sense to replicate it. Previously, the Enter key value would be added to the master password box, so there's a small possibility this could break functionality for users who may depend on it. I'm betting that number of users is very small if non-existent.

Minor additions include bumping up copyright dates and tweaking SVN ignore lists, as well as a few clean-up items to make Lint for Android happy.