Changes to pass Google Play validation erros
Looks like Google Play is changed how it handles icons in the APK
format. Previously, they let you use XML files as "pointers" or
"symbolic links" to point to a single PNG file so you didn't have to
bloat the APK with multiple versions of the same file. This seems to
work fine with other resources, but not the main application launcher
icon. So I had to replace the XML "pointers" with the copies of the PNGs
they pointed to.