Recognize ES File Manager Pro and Total Commander
Adding support for ES File Manager Pro and Total Commander file
managers. Both of these are pretty easy to add. ESFMPro is simply the
same as the free version, only the package name is slightly different,
so we just need to add that as a recognized value. The same basically
goes for Total Commander, which supports the OpenIntents intent options;
there we can reuse the OI File Manager code. The only problem I can see
is what happens if the user has both apps installed (i.e., both OI File
Manager and Total Commander, or both versions of ESFM). Newer versions
of Android let you pick preferred or "default" apps, and in testing it
seemed to honor my choice.

I need to get someone with ESFMPro to test this before pushing it to
Google Play. I gave up on ESFM a long time ago, so I'm not exactly
thrilled at the prospect of buying the Pro version just to test this
change.